requirements.txt
asgiref==3.8.1
certifi==2024.8.30
cffi==1.17.0
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
coverage==7.6.1
cryptography==43.0.0
Django==5.1
django-filter==24.3
django-rename-app==0.1.7
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
idna==3.8
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.4
jwt==1.3.1
MarkupSafe==2.1.5
packaging==24.1
pillow==10.4.0
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==2.9.0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.2
requests==2.32.3
setuptools==74.1.2
sqlparse==0.5.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.2

stripe~=10.10.0
forex-python~=1.8